Create lazy_propagation.cpp
"Lazy" saves a value which were not added to current node before this time to save time. Before working with current node we should update the node with its "lazy".  And after that we should start working with its node. 

To update current node, we should find an appropritate segment and add to its node the value. But we should send new value to its child's "lazy" nodes.